+++ share/pot/prepare.sh
@@ -4,7 +4,8 @@
 # shellcheck disable=SC2039
 prepare-help()
 {
-	echo "pot prepare [-hvS] -p pot -U URL -t tag -a aID -n potname -c cmd [-e port]"
+	echo "pot prepare [-hvS] -p pot -U URL -t tag -a aID -n potname -c cmd"
+	echo '        [-e port] [-N network-type] [-i ipaddr]'
 	echo '  -h print this help'
 	echo '  -h verbose'
 	echo '  -p pot : the pot image'
@@ -13,6 +14,8 @@ prepare-help()
 	echo '  -a aID : the allocation ID'
 	echo '  -n potname : the new potname (used instead of pot_tag)'
 	echo '  -c cmd : the command line to start the container'
+	echo '  -N network-type : new network type of the imported pot'
+	echo '  -i ipaddr : an ip address or the keyword auto (if applicable)'
 	echo '  -e port : the tcp port'
 	echo '            This option can be repeated multiple time, to export more ports'
 	echo '  -S : start immediately the newly generated pot'
@@ -21,12 +24,14 @@ prepare-help()
 pot-prepare()
 {
 	# shellcheck disable=SC2039
-	local _pname _o _URL _tag _tpname _cmd _ports _allocation_tag _new_pname _auto_start
+	local _pname _o _URL _tag _tpname _cmd _ports _allocation_tag _new_pname _auto_start _network_type _ipaddr
 	_pname=
 	_ports=
+	_network_type=
+	_ipaddr=
 	_auto_start="NO"
 	OPTIND=1
-	while getopts "hvp:U:t:c:e:a:n:S" _o ; do
+	while getopts "hvp:U:t:c:e:a:n:SN:i:" _o ; do
 		case "$_o" in
 		h)
 			prepare-help
@@ -67,6 +72,17 @@ pot-prepare()
 			;;
 		S)
 			_auto_start="YES"
+			;;
+		N)
+			if ! _is_in_list "$OPTARG" $_POT_NETWORK_TYPES ; then
+				_error "Network type $OPTARG not recognized"
+				clone-help
+				${EXIT} 1
+			fi
+			_network_type="$OPTARG"
+			;;
+		i)
+			_ipaddr=$OPTARG
 			;;
 		*)
 			prepare-help
@@ -115,7 +131,14 @@ pot-prepare()
 	else
 		_debug "pot $_imported_pname already imported - reusing it"
 	fi
-	if ! pot-cmd clone -P "${_imported_pname}" -p "${_new_pname}" -i auto ; then
+	_clone_network_opt=
+	if [ -n "$_network_type" ]; then
+		_clone_network_opt="-N $_network_type"
+	fi
+	if [ -n "$_ipaddr" ]; then
+		_clone_network_opt="$_clone_network_opt -i $_ipaddr"
+	fi
+	if ! pot-cmd clone -P "${_imported_pname}" -p "${_new_pname}" $_clone_network_opt ; then
 		_error "Not able to clone imported pot as $_new_pname"
 	fi
 	if [ -n "$_cmd" ]; then